Separate lifecycle tests into separate files
