+++ test/_util/helpers.js
@@ -223,3 +223,11 @@ export function sortAttributes(html) {
 		return '<' + pre + list.join('') + after;
 	});
 }
+
+
+export const spyAll = obj => Object.keys(obj).forEach( key => sinon.spy(obj,key) );
+export const resetAllSpies = obj => Object.keys(obj).forEach( key => {
+	if (obj[key].args) {
+		obj[key].resetHistory();
+	}
+});
